# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ReWear API Documentation",
        "description": "ReWear - Sustainable Fashion Marketplace API\n *\n * **Mission:** Making fashion sustainable and accessible\n *\n * **Tagline:** Buy, Sell, Donate\n *\n * ## Features\n * - 🛍️ **Marketplace:** Buy and sell pre-loved clothing\n * - 💚 **Donations:** Donate items to verified charities\n * - 🚚 **Delivery:** Local driver network\n * - 🌍 **Impact:** Track environmental and social impact\n *\n * ## Authentication\n * All protected endpoints require a JWT Bearer token in the Authorization header:\n * ```\n * Authorization: Bearer YOUR_JWT_TOKEN\n * ```\n *\n * ## Rate Limiting\n * - **Registration OTP:** 5 requests per email per hour, 5 per IP per hour\n * - **Login:** Standard Laravel rate limiting\n * - **API Calls:** 60 requests per minute\n *\n * ## User Roles\n * - **user:** Regular buyer/seller (default)\n * - **charity:** Organization receiving donations (admin-created)\n * - **driver:** Verified delivery personnel (admin-approved)\n * - **admin:** Platform administrator\n *\n * ## Support\n * - **Email:** support@rewear.com\n * - **Phone:** +961 70 123456\n * ",
        "contact": {
            "name": "ReWear Support Team",
            "email": "support@rewear.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
...
        },
...
    ],
    "paths": {
...
        "/api/auth/refresh-token": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh access token",
                "description": "Refresh access token using refresh token",
                "operationId": "6c58dfc55f9dedcdba92125730675c22",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "refresh_token"
                                ],
                                "properties": {
                                    "refresh_token": {
                                        "type": "string",
                                        "example": "abc123..."
                                    },
                                    "rotate": {
                                        "description": "Whether to rotate refresh token",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully"
                    },
                    "401": {
                        "description": "Invalid or expired refresh token"
                    }
                }
            }
        },
        "/api/auth/validate": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Validate JWT token",
                "description": "Validate JWT access token",
                "operationId": "05d0f42af604c105f19db78cfe36eecd",
                "responses": {
                    "200": {
                        "description": "Token is valid"
                    },
                    "401": {
                        "description": "Token is invalid or expired"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout from current device",
                "description": "Logout from current device (revoke refresh token)",
                "operationId": "3349894bd6fbca13950dce2d17e1791f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "refresh_token"
                                ],
                                "properties": {
                                    "refresh_token": {
                                        "type": "string",
                                        "example": "abc123..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout-all": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout from all devices",
                "description": "Logout from all devices (revoke all refresh tokens)",
                "operationId": "400eb29db9267b1f1f7685f7a84a5566",
                "responses": {
                    "200": {
                        "description": "Logged out from all devices"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/sessions": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get all active sessions",
                "description": "Get all active sessions (devices)",
                "operationId": "b476b67c08fcfc7de341986d5d82f798",
                "responses": {
                    "200": {
                        "description": "Sessions retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/token-stats": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get token statistics",
                "description": "Get token statistics for current user",
                "operationId": "aeccb290612fa96f1ef2d900e418471b",
                "responses": {
                    "200": {
                        "description": "Statistics retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Enter JWT token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearer": []
        }
    ]
}